Move the DELETE endpoint into web middleware
