+++ app/Http/Controllers/Api/PostsController.php
@@ -83,21 +83,4 @@ class PostsController extends ApiController
             }
         }
     }
-
-    /**
-     * Delete a resource in storage
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($postId)
-    {
-        $postDeleted = $this->posts->destroy($postId);
-
-        if ($postDeleted) {
-            return response()->json(['code' => 200, 'message' => 'Post deleted.']);
-        } else {
-            return response()->json(['code' => 500, 'message' => 'There was an error deleting the post']);
-        }
-    }
 }
